pyproject.toml
[project]
name = "yt-to-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
    "youtube-transcript-api>=1.1.0",
]
